zera-basemodule/basemodule.cpp,h: module nr becomes 16bit now, to increase number of possible modules in case of error calculators

+++ zera-basemodule/basemodule.cpp
@@ -202,7 +202,7 @@ void cBaseModule::stopModule()
 }
 
 
-quint8 cBaseModule::getModuleNr()
+quint16 cBaseModule::getModuleNr()
 {
     return m_nModuleNr;
 }
